requirements.txt
fastapi
uvicorn
python-multipart
requests
spacy
vi-core-news-lg
python-dotenv